# DM/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /employees:
        get:
            tags:
                - EmployeeService
            operationId: EmployeeService_GetAllEmployee
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetAllEmployeeRespone'
        post:
            tags:
                - EmployeeService
            operationId: EmployeeService_CreateEmployee
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.CreateEmployeeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.CreateEmployeeRespone'
    /employees/page:
        get:
            tags:
                - EmployeeService
            operationId: EmployeeService_GetEmployeeByPage
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetEmployeeByPageRespone'
    /employees/{id}:
        get:
            tags:
                - EmployeeService
            operationId: EmployeeService_GetEmployeeByID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetEmployeeByIDRespone'
        put:
            tags:
                - EmployeeService
            operationId: EmployeeService_UpdateEmployee
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.UpdateEmployeeResquest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.UpdateEmployeeRespone'
        delete:
            tags:
                - EmployeeService
            operationId: EmployeeService_DeleteEmployee
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.DeleteEmployeeRespone'
...
components:
    schemas:
...
        helloworld.v1.CreateEmployeeRequest:
            type: object
            properties:
                name:
                    type: string
                phone:
                    type: string
                password:
                    type: string
                role:
                    type: string
                salary:
                    type: integer
                    format: uint32
                subDepartmentId:
                    type: integer
                    format: uint32
        helloworld.v1.CreateEmployeeRespone:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
...
        helloworld.v1.DeleteEmployeeRespone:
            type: object
            properties: {}
...
        helloworld.v1.Employee:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                phone:
                    type: string
                password:
                    type: string
                salary:
                    type: integer
                    format: uint32
                role:
                    type: string
                subDepartmentId:
                    type: integer
                    format: uint32
...
        helloworld.v1.GetAllEmployeeRespone:
            type: object
            properties:
                employees:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.Employee'
...
        helloworld.v1.GetEmployeeByIDRespone:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                phone:
                    type: string
                password:
                    type: string
                role:
                    type: string
                salary:
                    type: integer
                    format: uint32
                subDepartmentId:
                    type: integer
                    format: uint32
        helloworld.v1.GetEmployeeByPageRespone:
            type: object
            properties:
                employees:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.Employee'
...
        helloworld.v1.UpdateEmployeeRespone:
            type: object
            properties: {}
        helloworld.v1.UpdateEmployeeResquest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                phone:
                    type: string
                password:
                    type: string
                role:
                    type: string
                salary:
                    type: integer
                    format: uint32
                subDepartmentId:
                    type: integer
                    format: uint32
...